pyproject.toml
[tool.poetry]
name = "ai-task-manager"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = { extras = ["standard"], version = ">=0.115.11,<0.116.0" }
sqlmodel = "^0.0.23"
pyjwt = "^2.10.1"
aiosqlite = "^0.21.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.28.1"
tenacity = "^9.0.0"
langchain-google-genai = "^2.1.0"
langgraph = "^0.3.8"
langchain = "^0.3.20"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
